requirements/base.txt
# Base
python-slugify==8.0.1
pillow==10.1.0
psycopg==3.1.14 --no-binary psycopg

# Django
django>=5.0.0

# Environment
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
django-filter==23.4
django-cors-headers == 4.3.1

# JWT
pyjwt==2.8.0

# Passwords security
argon2-cffi==23.1.0

# Static files
whitenoise==6.6.0

# Celery
redis==5.0.1
django-redis==5.4.0

psycopg2-binary
dj-database-url
whitenoise[brotli]
gunicorn 
uvicorn